+++ eslint.config.js
@@ -1,12 +1,13 @@
 import {defineConfig, globalIgnores} from 'eslint/config';
 import globals from 'globals';
 
-import importConfig from './src/configs/import-order.js';
-import prettierConfig from './src/configs/prettier.js';
-import mainConfig from './src/index.js';
+import importConfig from './import-order.js';
+import prettierConfig from './prettier.js';
+
+import mainConfig from './index.js';
 
 export default defineConfig(
-    globalIgnores(['lib/**/*'], 'Ignore generated files'),
+    globalIgnores(['types/**/*'], 'Ignore generated files'),
     mainConfig,
     importConfig,
     prettierConfig,
