+++ index.js
@@ -1,0 +1,6 @@
+import baseConfig from './base.js';
+import typescriptConfig from './typescript.js';
+
+const mainConfig = [...baseConfig, ...typescriptConfig];
+
+export default mainConfig;
