--- node-helpers/register.js
@@ -1,3 +1,0 @@
-import {register} from 'node:module';
-
-register('./ts-extension-override.js', import.meta.url);
