--- src/index.ts
@@ -1,6 +1,0 @@
-import baseConfig from './configs/base.js';
-import typescriptConfig from './configs/typescript.js';
-
-const mainConfig = [...baseConfig, ...typescriptConfig];
-
-export default mainConfig;
