+++ tsconfig.publish.json
@@ -1,8 +1,7 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["./src"],
+  "include": ["./*.js"],
   "compilerOptions": {
-    "rootDir": "src",
     "declaration": true
   }
 }
